Add tests for concurrent output and permissions
