+++ common/tests/scripts/slow-output.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+a=0
+while true; do
+    echo "line $a"
+    a=$(($a+1))
+    sleep 1
+done
